refactor: SurveyCategory 구조 단순화
