+++ module-domain/src/main/kotlin/org/depromeet/team3/surveycategory/SurveyCategory.kt
@@ -1,6 +1,7 @@
 package org.depromeet.team3.surveycategory
 
 import org.depromeet.team3.common.BaseTimeDomain
+import org.depromeet.team3.common.enums.SurveyCategoryType
 import java.time.LocalDateTime
 
 data class SurveyCategory (
