+++ module-domain/src/main/kotlin/org/depromeet/team3/surveycategory/SurveyCategoryRepository.kt
@@ -1,5 +1,7 @@
 package org.depromeet.team3.surveycategory
 
+import org.depromeet.team3.common.enums.SurveyCategoryType
+
 interface SurveyCategoryRepository {
 
     fun save(surveyCategory: SurveyCategory): SurveyCategory
@@ -17,4 +19,6 @@ interface SurveyCategoryRepository {
     fun existsBySortOrderAndParentIdAndIsDeletedFalse(sortOrder: Int, parentId: Long?, excludeId: Long? = null): Boolean
     
     fun countChildrenByParentIdAndIsDeletedFalse(parentId: Long): Long
+    
+    fun findByNameAndType(name: String, type: SurveyCategoryType): SurveyCategory?
 }
