--- module-domain/src/main/kotlin/org/depromeet/team3/surveycategory/SurveyCategoryType.kt
@@ -1,8 +1,0 @@
-package org.depromeet.team3.surveycategory
-
-enum class SurveyCategoryType {
-    CUISINE,           // 선호하는 음식
-    AVOID_INGREDIENT,  // 피해야 하는 재료
-    AVOID_MENU         // 원하지 않는 메뉴
-}
-
