+++ module-infra/src/main/kotlin/org/depromeet/team3/surveycategory/SurveyCategoryEntity.kt
@@ -2,6 +2,7 @@ package org.depromeet.team3.surveycategory
 
 import jakarta.persistence.*
 import org.depromeet.team3.common.BaseTimeEntity
+import org.depromeet.team3.common.enums.SurveyCategoryType
 
 @Entity
 @Table(
